moving storage_information stream to separate file
